+++ src/errors.rs
@@ -22,6 +22,8 @@ pub enum CropBoxError {
     PositionIsOutOfImageBoundaries,
     #[error("Size of the crop box is out of the image boundaries")]
     SizeIsOutOfImageBoundaries,
+    #[error("Width or height of the crop box is less or equal to zero")]
+    WidthOrHeightLessOrEqualToZero,
 }
 
 #[derive(Error, Debug, Clone, Copy)]